[Issue-3512] Refactor code
